chore: Move data into state/info, fix imports
