+++ src/components/InfoTables/PoolsTable.tsx
@@ -3,8 +3,8 @@ import styled from 'styled-components'
 import { Link } from 'react-router-dom'
 import { Text, Flex, Box, Skeleton, ArrowBackIcon, ArrowForwardIcon } from '@pancakeswap/uikit'
 import { formatAmount } from 'utils/formatInfoNumbers'
-import { PoolData } from 'state/pools/types'
-import { ITEMS_PER_INFO_TABLE_PAGE } from 'config/info'
+import { PoolData } from 'state/info/types'
+import { ITEMS_PER_INFO_TABLE_PAGE } from 'config/constants/info'
 import { DoubleCurrencyLogo } from 'components/CurrencyLogo'
 import { useTranslation } from 'contexts/Localization'
 import { ClickableColumnHeader, TableWrapper, PageButtons, Arrow, Break } from './shared'
