+++ src/components/TopTokenMovers/index.tsx
@@ -2,8 +2,8 @@ import React, { useMemo, useRef, useState, useEffect } from 'react'
 import styled from 'styled-components'
 import { Text, Flex, Box, Card } from '@pancakeswap/uikit'
 import { Link } from 'react-router-dom'
-import { useAllTokenData } from 'state/tokens/hooks'
-import { TokenData } from 'state/tokens/types'
+import { useAllTokenData } from 'state/info/hooks'
+import { TokenData } from 'state/info/types'
 import { CurrencyLogo } from 'components/CurrencyLogo'
 import { formatAmount } from 'utils/formatInfoNumbers'
 import Percent from 'components/Percent'
