+++ src/state/info/hooks.ts
@@ -4,12 +4,12 @@ import { getUnixTime, startOfHour, Duration, sub } from 'date-fns'
 import { AppState, AppDispatch } from 'state'
 import { isAddress } from 'utils'
 import { ChartDayData, PriceChartEntry, Transaction } from 'types'
-import fetchPoolChartData from 'data/pools/chartData'
-import fetchPoolTransactions from 'data/pools/transactions'
-import fetchTokenChartData from 'data/tokens/chartData'
-import fetchTokenTransactions from 'data/tokens/transactions'
-import { fetchTokenPriceData } from 'data/tokens/priceData'
-import { fetchPoolsForToken } from 'data/tokens/poolsForToken'
+import fetchPoolChartData from 'state/info/queries/pools/chartData'
+import fetchPoolTransactions from 'state/info/queries/pools/transactions'
+import fetchTokenChartData from 'state/info/queries/tokens/chartData'
+import fetchTokenTransactions from 'state/info/queries/tokens/transactions'
+import { fetchTokenPriceData } from 'state/info/queries/tokens/priceData'
+import { fetchPoolsForToken } from 'state/info/queries/tokens/poolsForToken'
 import { notEmpty } from 'utils/infoUtils'
 import {
   updateProtocolData,
